# cmd/root2npy: improve memory usage